Initial commit of my Slowly letter downloader
I put this together over the past couple of days. It's incredibly
rough around the edges, and has many issues. However, today I was able
to download all letters from a contact on Slowly. This included text,
stamp, photos, pretty much everything.

I intend to refine this code, and to also release a "portable" version
of the code which would be easier for people to access.