Metadata now embedded into PDFs to skip duplicates
Each PDF that is created is now embedded with data that stores which
number letter it is, and the name of the penpal chosen by the user.
Using this data the code can now skip over existing PDFs, even if
they've been renamed.

This also means if the code unexpectedly crashes or a users internet
cuts out, it's simple enough to load up the code again and continue
from where you left off, without having to redo all the ones you've
already downloaded.