Further work on GUI tests
Buttons have been changed to checkboxes, which is more suitable for the
purpose of the app. They are currently locked to the centre, whether or
not I want them to stay like this, I'm not sure.

A scrollbar has also been added to the left frame, so a large list of
penpals can be stored and scrolled through by the user.