Imported GUI interface into main.py
A new module has been created, interface.py, which contains all the
code for the GUI. This is imported into main.py and takes one argument
upon inialization, which is currently a list containing the names of
all penpals that have been written to by the user.

Currently there's nothing that can be done with the GUI, it is purely
a proof of concept for now.