Tests to integrate browser into app
After much trial an error, there is now an embedded browser in the
interface-embeded-web.py GUI test file. I don't know how stable this
browser will be, but it'll likely serve a singular purpose: to allow
the user to login, save the session in the sessions folder, and then
remove itself to be replaced by a progress bar.

Chrome will run in a headless mode in the background performing the
required tasks, and will only update the user via the progress bar,
as this is all they need to see.