Now have a fully functional GUI!
The GUI is now fully integrated with the underlying selenium code,
allowing the user to choose which penpals they wish to download the
letters from.

The main issue right now is that Chrome (selenium) refuses to download
any PDFs whilst in headless mode, which is what I wanted. I suspect
this might have something to do with my editing of the selenium code
to make selenium download PDFs without headless mode.