Programs basic functions all work in headless mode
The custom version of Selenium is no longer required, so this has been
removed. A new process for printing PDF files has been implemented,
and whilst I prefer the format of the older one, I think this method
will allow for more customization, so in the long run, this is better.

I'll know by researching how to use Pyinstaller so I can release either
a standalone executable, or a zip file containing an EXE with a few
subdirectories for data.