Setup.py and expanded README.MD
Numerous improvements to the underlying code, mostly to make it
compatible with cx-Freeze, which is used to create the Windows binary.
Setup.py will successfully create an executable Windows binary file,
there are however still a few bugs to sort out before an alpha release.

README.MD has been expanded, informing the user on things such as
current and planned features, how to use the program, etc. No doubt as
development progresses this will need to be further updated, but this
is a decent starting point.