Bug fixes
Managed to fix a bug caused by selenium trying to show the download
progress bar once the program was compiled. This was fixed by
commenting out code from the selenium module, so I might make some code
to automate this process in the future.

Included fail safes for the junction creation process, by giving a
copy method as a backup. This code is perhaps a bit messy right now,
but will be cleaned up at some point. Importantly, the compiled exe now
fully works on the alpha testers device, and in my virtual test env.