GUI improvements and initial alpha binary release
Various fixes and enhancements implemented to create a more pleasing
GUI, such as removing blank white spaces and giving an overall more
consistent feel. An icon has also been made by the program, I'm no
graphic designer, so... Yeah.

The current version seems to be functional enough for a 0.1 alpha
release. I'm nervous as there are still some issues of chromedriver
persisting after program exit, and the program is still really slow.
But, I figure it's time to open myself to strangers feedback on this
project.