The threading module has been added!
It is however very buggy, with a lot of functions seeming to currently
not function entirely as intended. A lot of functions are going to have
to be reworked to "play nice" with this addition.

A few issues have been closed with this commit. #17 is closed with the
addition of the settings button in the app, as bare bones as it might
be. Also #16 is closed too.