Fixed threading issues + chrome download prog bar
An issue that prevented the application from running properly which
occurred around the time I introduced the threading module has been
fixed.

Also finally the chrome download progress bar has been reimplemented
after being absent for a very long time. It no longer slows down the
application and by extension the download. It's not much of a bar, it's
actually just text, but it does the job until the bar is reintroduced.