Goodbye progress bars, hello progress circles!!!
Progress circles have been added to a few different frames in the
program. Those being the download frame for when Chrome is being
downloaded, and when letters from a penpal are being printed.

The implementation isn't perfect, it has a tendency to speed up if not
properly unloaded before being loaded again. This closes #21.