Significant speedups
Removed a few instances of time.sleep and also reduced wait time in
remaining instances of time.sleep. May have reduced them a bit too
ambitiously, but you don't know until you try, so here we go. This
makes significant progress for issue #4, but I hesitate to close it as
I can probably optimise further.

Also, the cefpython integrated browser is now resizable. I'd like to
eventually remove cefpython, but for now it's the method of getting
the cache files that I'm aware of. Now I think of it, I'm going to
open an issue about this to explore my ideas further.