0.2 Release!!!
This commit fixes issue #34, this was done by creating a new function
which checks for the current version of the portable Chrome release,
regardless of whether or not this is being used. This is done so the
ChromeDriverManager (which didn't want to find the latest version on
its own) can be instructed to download the current stable version.

After some extensive testing on my newly accquired laptop, and in my
usual testing environment I feel confident in finally releasing version
0.2, bringing a huge amount of improvements and fixes. Yay!